scheduled notifications tests + settings info

+++ screens/menu/PrescriptionScreen.js
@@ -4,6 +4,7 @@ import { View } from 'react-native';
 import { AuthContext } from '../../context/AuthContext';
 import PrescriptionService from '../../services/prescription.service';
 import PrescriptionItem from '../../components/PrescriptionItem';
+import * as Network from 'expo-network';
 
 const db = SQLite.openDatabase('db.prescriptionDb');
 const PrescriptionScreen = () => {
@@ -11,6 +12,7 @@ const PrescriptionScreen = () => {
     const { username } = useContext(AuthContext);
 
     useEffect(() => {
+
         db.transaction(tx => {
             tx.executeSql(
                 'CREATE TABLE IF NOT EXISTS prescription_items ( \
@@ -45,48 +47,52 @@ const PrescriptionScreen = () => {
         fetchData();
     }, []);
 
-    const fetchData = () => {
-        PrescriptionService.getPrescription(username).then(response => {
-            console.log(response.data);
-            response.data.prescriptionItems.forEach(item => {
-                db.transaction((tx) => {
-                    tx.executeSql("DELETE FROM items_times_of_day");
-                    tx.executeSql("DELETE FROM prescription_items");
-                    tx.executeSql('INSERT INTO prescription_items \
+    const fetchData = async () => {
+        const networkStatus = await Network.getNetworkStateAsync();
+        if (networkStatus.isInternetReachable) {
+            PrescriptionService.getPrescription(username).then(response => {
+                console.log(response.data);
+                response.data.prescriptionItems.forEach(item => {
+                    db.transaction((tx) => {
+                        tx.executeSql("DELETE FROM items_times_of_day");
+                        tx.executeSql("DELETE FROM prescription_items");
+                        tx.executeSql('INSERT INTO prescription_items \
                     (medicament, quantity, intake_interval, available_until) \
                     values (?, ?, ?, ?)',
-                        [item.medicament.name, item.quantity, item.daysBetweenAdministrations, item.administrationEndDate
-                        ],
-                        (_, resultSet) => {
-                            item.id = resultSet.insertId;
-                            item.timesOfDay.forEach(time => {
-                                switch (time.name) {
-                                    case "MORNING":
-                                        tx.executeSql("INSERT INTO items_times_of_day (item_id, time_of_day_id) \
+                            [item.medicament.name, item.quantity, item.daysBetweenAdministrations, item.administrationEndDate
+                            ],
+                            (_, resultSet) => {
+                                item.id = resultSet.insertId;
+                                item.timesOfDay.forEach(time => {
+                                    switch (time.name) {
+                                        case "MORNING":
+                                            tx.executeSql("INSERT INTO items_times_of_day (item_id, time_of_day_id) \
                                         values (?, ?)", [resultSet.insertId, 1]);
-                                        break;
-                                    case "NOON":
-                                        tx.executeSql("INSERT INTO items_times_of_day (item_id, time_of_day_id) \
+                                            break;
+                                        case "NOON":
+                                            tx.executeSql("INSERT INTO items_times_of_day (item_id, time_of_day_id) \
                                         values (?, ?)", [resultSet.insertId, 2]);
-                                        break;
-                                    case "EVENING":
-                                        tx.executeSql("INSERT INTO items_times_of_day (item_id, time_of_day_id) \
+                                            break;
+                                        case "EVENING":
+                                            tx.executeSql("INSERT INTO items_times_of_day (item_id, time_of_day_id) \
                                         values (?, ?)", [resultSet.insertId, 3]);
-                                        break;
-                                }
-                            })
-                        },
-                        (_, error) => console.log(error));
+                                            break;
+                                    }
+                                })
+                            },
+                            (_, error) => console.log(error));
 
-                })
+                    })
+                });
             });
-            db.transaction((tx) => {
-                // tx.executeSql('SELECT * FROM items_times_of_day', null,
-                //     (_, { rows: { _array } }) => console.log(_array));
-                // tx.executeSql('SELECT * FROM times_of_day', null,
-                //     (_, { rows: { _array } }) => console.log(_array));
+        }
+        db.transaction((tx) => {
+            // tx.executeSql('SELECT * FROM items_times_of_day', null,
+            //     (_, { rows: { _array } }) => console.log(_array));
+            // tx.executeSql('SELECT * FROM times_of_day', null,
+            //     (_, { rows: { _array } }) => console.log(_array));
 
-                tx.executeSql('SELECT \
+            tx.executeSql('SELECT \
                     prescription_items.medicament, \
                     prescription_items.quantity, \
                     prescription_items.intake_interval, \
@@ -97,15 +103,14 @@ const PrescriptionScreen = () => {
                      INNER JOIN items_times_of_day ON items_times_of_day.item_id = prescription_items.id \
                      INNER JOIN times_of_day ON items_times_of_day.time_of_day_id = times_of_day.id \
                      ', null,
-                    (_, { rows: { _array } }) => {
-                        console.log(_array);
-                        setData(_array);
-                    },
-                    (_, error) => console.log("eroare: " + error));
-            });
+                (_, { rows: { _array } }) => {
+                    console.log(_array);
+                    setData(_array);
+                },
+                (_, error) => console.log("eroare: " + error));
         });
-
     }
+
 
     return (
         <View>
@@ -120,7 +125,7 @@ const PrescriptionScreen = () => {
                     })
             }
         </View>
-    )
+    );
 }
 
 export default PrescriptionScreen;
